DataViews Extensibility: Allow unregistering the duplicate post action
