+++ packages/editor/src/dataviews/store/private-actions.ts
@@ -21,6 +21,7 @@ import restorePost from '../actions/restore-post';
 import type { PostType } from '../types';
 import { store as editorStore } from '../../store';
 import { unlock } from '../../lock-unlock';
+import duplicatePost from '../actions/duplicate-post';
 
 export function registerEntityAction< Item >(
 	kind: string,
@@ -87,6 +88,14 @@ export const registerPostTypeActions =
 			.getCurrentTheme();
 
 		const actions = [
+			// @ts-ignore
+			globalThis.IS_GUTENBERG_PLUGIN
+				? ! [ 'wp_template', 'wp_block', 'wp_template_part' ].includes(
+						postTypeConfig.slug
+				  ) &&
+				  canCreate &&
+				  duplicatePost
+				: undefined,
 			postTypeConfig.slug === 'wp_template_part' &&
 				canCreate &&
 				currentTheme?.is_block_theme &&
