+++ packages/editor/src/dataviews/types.ts
@@ -14,6 +14,19 @@ export interface BasePost {
 	type: string;
 	id: string | number;
 	blocks?: Object[];
+
+	// Not sure if these should be in template, template part or pattern types
+	comment_status?: 'open' | 'closed';
+	excerpt?: string | { raw: string; rendered: string };
+	meta?: Record< string, any >;
+	parent?: number;
+	password?: string;
+	template?: string;
+	format?: string;
+	featured_media?: number;
+	menu_order?: number;
+	ping_status?: 'open' | 'closed';
+	_links?: Record< string, { href: string }[] >;
 }
 
 export interface Template extends BasePost {
@@ -36,10 +49,6 @@ export interface Pattern extends BasePost {
 	slug: string;
 	title: { raw: string };
 	wp_pattern_sync_status: string;
-}
-
-export interface PostWithPageAttributesSupport extends BasePost {
-	menu_order: number;
 }
 
 export type Post = Template | TemplatePart | Pattern | BasePost;
